requirements.txt
appnope==0.1.3
asgiref==3.7.0
asttokens==2.2.1
backcall==0.2.0
black==23.7.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.6
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
Django==4.2.2
django-elasticsearch-dsl==7.3
django-rest-framework==0.1.0
djangorestframework==3.14.0
dnspython==2.3.0
elasticsearch==7.13.4
elasticsearch-dsl==7.4.1
executing==1.2.0
idna==3.4
ipykernel==6.23.1
ipython==8.14.0
isort==5.12.0
jedi==0.18.2
jupyter_client==8.2.0
jupyter_core==5.3.0
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
nest-asyncio==1.5.6
packaging==23.1
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.5.1
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg==3.1.9
psycopg2==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pymongo==3.11.4
python-dateutil==2.8.2
pytz==2023.3
pyzmq==25.1.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.6.3
urllib3==1.26.16
wcwidth==0.2.6
